Move local attack to new file
